# G-times speedup and G-times memory reduction in event_based calculations