--- a/openquake/hazardlib/contexts.py
+++ b/openquake/hazardlib/contexts.py
@@ -816,7 +816,6 @@ class ContextMaker(object):
                     if small_distances.any():
                         array = numpy.array(array)  # make a copy first
                         array[small_distances] = self.minimum_distance
-                        array.flags.writeable = False
                         ctx[name] = array
             slc = slice(start, start + len(ctx))
             for par in dd:
@@ -1158,7 +1157,6 @@ class ContextMaker(object):
         :param rup_indep: rupture flag (false for mutex ruptures)
         :yields: poes, ctxt, invs with poes of shape (N, L, G)
         """
-        ctx.flags.writeable = True
         ctx.mag = numpy.round(ctx.mag, 3)
         for mag in numpy.unique(ctx.mag):
             ctxt = ctx[ctx.mag == mag]
@@ -1216,7 +1214,6 @@ class ContextMaker(object):
         for ctx in ctxs:
             for poes, ctxt, invs in self.gen_poes(ctx, rup_indep):
                 with self.pne_mon:
-                    ctxt.flags.writeable = True  # avoid numba type error
                     pmap.update(poes, invs, ctxt, itime, rup_mutex)
 
     # called by gen_poes and by the GmfComputer
@@ -1245,8 +1242,6 @@ class ContextMaker(object):
             start = 0
             for ctx in recarrays:
                 slc = slice(start, start + len(ctx))
-                # make the context immutable
-                ctx.flags.writeable = False
                 adj = compute(gsim, ctx, self.imts, *out[:, g, :, slc])
                 if adj is not None:
                     self.adj[gsim].append(adj)
@@ -1273,8 +1268,6 @@ class ContextMaker(object):
         start = 0
         for ctx in recarrays:
             slc = slice(start, start + len(ctx))
-            # make the context immutable
-            ctx.flags.writeable = False
             adj = compute(gsim, ctx, self.imts, *out[:, :, slc])
             if adj is not None:
                 self.adj[gsim].append(adj)