… openquake/hazardlib/tests/site_test.py