G is the number of GSIMs, G=19 for EUR so a huge memory reduction is expecting, but this is visible only if you have millions of sites. There is also a huge speedup, ideally of G times, visible also with few sites:
```
# EUR, 13x speedup in " computing mean_stds"
| calc_510, maxmem=44.4 GB   | time_sec | memory_mb | counts  |
|----------------------------+----------+-----------+---------|
| computing mean_stds        | 9_236    | 0.0       | 261_700 |
| computing mean_stds        | 717.8    | 0.0       | 284_334 |
```
Here are the numbers for SAM, ~20,000 ruptures, half million sites and G=3:
```
# before
| calc_2804, maxmem=6.1 GB   | time_sec | memory_mb | counts |
|----------------------------+----------+-----------+--------|
| total event_based          | 427.8    | 45.2305   | 165    |
| computing mean_stds        | 170.2519 | 0.0       | 20_257 |
| instantiating GmfComputer  | 127.0037 | 0.0       | 20_381 |
| computing gmfs             | 94.6061  | 0.0       | 20_342 |
# after
| calc_2803, maxmem=12.0 GB  | time_sec | memory_mb | counts |
|----------------------------+----------+-----------+--------|
| total event_based          | 318.5    | 65.0625   | 165    |
| computing mean_stds        | 58.1481  | 0.0       | 20_342 |
| instantiating GmfComputer  | 127.9250 | 0.0       | 20_381 |
| computing gmfs             | 96.3396  | 0.0       | 20_342 |
```
Notice the 2x reduction in memory and the 3x speedup in "computing mean_stds" due to wasted calculations in the previous version ( the speedup in due to `numpy.isin(self.rlz, rlzs)` being false most of the times).